requirements.txt
matplotlib==3.3.2
numpy==1.19.2
office31==0.1.3
scikit-build==0.11.1
scikit-image==0.17.2
scikit-learn==0.24.1
opencv-python==4.2.0.34
pandas==1.1.3
pathtools==0.1.2
pillow==8.0.0
pytorch-lightning==1.3.3
sklearn==0.0
torch==1.8.0	
torchmetrics==0.3.2
torchvision==0.9.0
wandb==0.10.32